java: realtime timeseries db integration w/ timescale

+++ java/src/main/java/com/leothenardo/homebroker/orders/application/EmitOrderServiceInputDTO.java
@@ -7,6 +7,6 @@ public record EmitOrderServiceInputDTO(
 				String assetId,
 				int shares,
 				OrderType type,
-				double price
+				Float price
 ) {
 }
